Add files via upload
I don't know how to maintain the directory structure of the project on my computer in the project after upload here.
meme.py and app.py are supposed to be in the top directory and the rest are to be in directory QuoteEngine.  References to DocxIngestor and docx are commented out because I couldn't get docx installed - to be figured out later.  
My immediate problem is described in my Mentor Forum post of Sept. 7, "Problem with Image.open in make_meme.py".